production and damage effects

+++ Assets/Scripts/Game/Damage/Damage.cs
@@ -6,11 +6,11 @@ namespace Game.Damage
     {
         [Flags] public enum Type { HealthLoss = 1 << 0, Physical = 1 << 1, Fire = 1 << 2 }
 
-        public int amount;
+        public float amount;
         public Type type;
         public IDamageSource source;
 
-        public Damage(int amount, Type type, IDamageSource source)
+        public Damage(float amount, Type type, IDamageSource source)
         {
             this.amount = amount;
             this.type = type;
